Feat : 천단위 콤마 구현
