+++ src/ViewModel.tsx
@@ -1,0 +1,21 @@
+import { useState } from "react";
+
+const ViewModel = () => {
+  const [isReady, setReady] = useState(false);
+  const [eurInfo, setEurInfo] = useState<any>({});
+
+  const getEurInfo = async () => {
+    const krweur = await fetch(
+      "https://quotation-api-cdn.dunamu.com/v1/forex/recent?codes=FRX.KRWEUR"
+    )
+      .then((response) => response.json())
+      .then((array) => array[0]);
+
+    setEurInfo(krweur);
+    setReady(true);
+  };
+
+  return { isReady, eurInfo, getEurInfo };
+};
+
+export default ViewModel;
